Separate capabilities resolving to custom class
